Fix bug in completion detection
As we iterate between constraining nodes based on valence, and
constraining edges, the intermediate state can be messy, since either
the valence doesn't match the edge possibilities, or the ends of an
edge don't agree. This makes it tricky to work out which edges to
traverse when trying to find a complete component. As such, the
completeness-checker now only traverses edges from complete islands,
and only peers into incomplete islands to check that this means the
current subgraph isn't a complete component.

The problem was found while testing a larger case. The output when no
solution was found was complete silence, so I've also changed it so
that if no solution was found it says so.